Created a New Visual Studio Project Called FrontRowCMS2 With Identity
As part of the preperation for milestone 4, I created a new project with Identity integration and am in the process of moving all the files from our current project to this new one. It is very important when the team implements their models to use FrontRowCMS2  as the original FrontRowCMS project will be deleted when I get approval from all team members.